server/src/requirements.txt
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-datastore==1.9.0
google-cloud-tasks==1.1.0
googleapis-common-protos==1.6.0
pyjwt==1.7.1
requests==2.22.0
marshmallow==3.2.1
python-dotenv==0.10.3
PyMySQL==0.9.3
PyYAML==5.1.2
SQLAlchemy==1.3.8
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
SQLAlchemy-Utils==0.34.2
Jinja2==2.10.3
GitPython==3.0.3
cryptography==2.7